backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
djongo==1.3.6
django-cors-headers==4.4.0
python-dateutil==2.9.0
sqlparse==0.2.4
pymongo==3.12.3
